% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel.R
\name{build_hankel_matrix}
\alias{build_hankel_matrix}
\title{Build Hankel-Takens Matrix}
\usage{
build_hankel_matrix(y, delays)
}
\arguments{
\item{y}{Numeric matrix (n_obs x n_time).}

\item{delays}{Number of delay steps.}
}
\value{
Hankel matrix of dimension ((delays + 1) * n_obs) x (n_time - delays).
}
\description{
Constructs the Hankel (Takens) matrix from time series data for
delay embedding.
}
\details{
The Hankel-Takens matrix has the structure (eq. 118 from Mezic 2020):
\deqn{H_{i,j} = y(:, i+j-1)}

For a scalar time series y = \link{y1, y2, ..., yn}, with d delays:
\deqn{H = \begin{bmatrix}
  y_1 & y_2 & \cdots & y_{n-d} \\
  y_2 & y_3 & \cdots & y_{n-d+1} \\
  \vdots & & & \vdots \\
  y_{d+1} & y_{d+2} & \cdots & y_n
\end{bmatrix}}
}
\keyword{internal}
