% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel.R
\name{compute_companion_matrix}
\alias{compute_companion_matrix}
\title{Compute Companion Matrix}
\usage{
compute_companion_matrix(H)
}
\arguments{
\item{H}{Hankel matrix.}
}
\value{
Companion matrix C such that H2 ≈ H1 C.
}
\description{
Computes the companion matrix representation for Krylov subspace DMD.
}
\details{
The companion matrix has the structure (eq. 83):
\deqn{C = \begin{bmatrix}
  0 & 0 & \cdots & 0 & c_1 \\
  1 & 0 & \cdots & 0 & c_2 \\
  0 & 1 & \cdots & 0 & c_3 \\
  \vdots & & \ddots & & \vdots \\
  0 & 0 & \cdots & 1 & c_n
\end{bmatrix}}
}
\keyword{internal}
