% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\title{Legacy DMD Functions (Deprecated)}
\description{
These functions are deprecated and will be removed in a future version.
Please use the new API instead.
}
\keyword{internal}
