% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{dmd_convergence}
\alias{dmd_convergence}
\title{Estimate Convergence Rate}
\usage{
dmd_convergence(X, sample_fractions = c(0.25, 0.5, 0.75, 1), rank = NULL, ...)
}
\arguments{
\item{X}{Data matrix (n_vars x n_time).}

\item{sample_fractions}{Numeric vector of fractions of data to use.
Default is \code{c(0.25, 0.5, 0.75, 1.0)}.}

\item{rank}{Rank for DMD (NULL for auto).}

\item{...}{Additional arguments passed to \code{\link[=dmd]{dmd()}}.}
}
\value{
A list containing:
\describe{
\item{sample_sizes}{Vector of sample sizes used}
\item{eigenvalues}{List of eigenvalue vectors at each sample size}
\item{eigenvalue_changes}{Vector of max eigenvalue changes between successive fits}
\item{convergence_estimate}{Estimated convergence rate}
}
}
\description{
Estimates the convergence rate of the DMD approximation by computing
eigenvalues at increasing sample sizes and measuring convergence.
}
\details{
From Theorem 4.3 in Mezic (2020), for systems with pure point spectrum,
the finite section approximation converges at rate O(1/m) where m is
the number of samples. This function empirically estimates this rate.
}
\examples{
t <- seq(0, 20, by = 0.1)
X <- rbind(cos(t), sin(t))
conv <- dmd_convergence(X)
cat("Estimated convergence rate:", conv$convergence_estimate, "\n")

}
