% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifting.R
\name{generate_monomials}
\alias{generate_monomials}
\title{Generate Monomial Exponents}
\usage{
generate_monomials(n_vars, max_degree)
}
\arguments{
\item{n_vars}{Number of variables.}

\item{max_degree}{Maximum total degree.}
}
\value{
Matrix where each row is an exponent vector.
}
\description{
Generates all combinations of exponents for monomials up to a given degree.
}
\keyword{internal}
