% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getAMatrix}
\alias{getAMatrix}
\title{Compute Koopman Matrix (Deprecated)}
\usage{
getAMatrix(data, p = 1, comp = NA)
}
\arguments{
\item{data}{The matrix data for which you wish to predict future columns
expressed as a 2D matrix or data frame.}

\item{p}{The percentage of explanation of the SVD required (default = 1).}

\item{comp}{The number of components of the SVD to keep (> 1).
Default is NULL, overrides p if set.}
}
\value{
An object of class "kdmd" containing the Koopman matrix.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Use \code{\link[=dmd]{dmd()}} instead, which returns
a richer object containing the A matrix along with modes, eigenvalues,
and other useful components.
}
\examples{
\dontrun{
# Old way (deprecated)
m <- matrix(1:30, nrow = 3)
A <- getAMatrix(m)

# New way
model <- dmd(m)
A <- model$A
}

}
\seealso{
\code{\link[=dmd]{dmd()}} for the new API.
}
\keyword{internal}
