% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gla.R
\name{gla_reconstruct}
\alias{gla_reconstruct}
\title{Reconstruct Signal Using GLA Model}
\usage{
gla_reconstruct(object, modes_to_use = NULL)
}
\arguments{
\item{object}{A \code{"gla"} object.}

\item{modes_to_use}{Integer vector; which modes to include.
Default is all modes.}
}
\value{
Matrix of reconstructed values (n_obs x n_time).
}
\description{
Reconstructs the original signal using the computed eigenfunctions
and modes.
}
