% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel.R
\name{hankel_reconstruct}
\alias{hankel_reconstruct}
\title{Reconstruct Time Series from Hankel-DMD}
\usage{
hankel_reconstruct(object, n_steps = NULL)
}
\arguments{
\item{object}{A \code{"hankel_dmd"} object.}

\item{n_steps}{Number of steps to reconstruct. Default is original length.}
}
\value{
Reconstructed time series matrix (n_obs x n_steps).
}
\description{
Reconstructs the original time series using the Hankel-DMD model.
}
