% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifting.R
\name{is_delay_lifting}
\alias{is_delay_lifting}
\title{Check if Lifting Specification is Delay-based}
\usage{
is_delay_lifting(lifting)
}
\arguments{
\item{lifting}{Lifting specification}
}
\value{
Logical; TRUE if delay-based
}
\description{
Determines if a lifting specification involves time-delay embedding,
which reduces the number of usable time points.
}
\keyword{internal}
