% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{kdmd}
\alias{kdmd}
\title{Create a kdmd Object (Deprecated)}
\usage{
kdmd(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
An object of class "kdmd".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Use \code{\link[=dmd]{dmd()}} instead.
}
\examples{
\dontrun{
# Old way (deprecated)
m <- matrix(1:20, nrow = 2)
A <- getAMatrix(m)

# New way
model <- dmd(m)
A <- model$A
}

}
\keyword{internal}
