% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifting.R
\name{lift_poly_cross}
\alias{lift_poly_cross}
\title{Polynomial Lifting with Cross-terms}
\usage{
lift_poly_cross(X, degree = 2)
}
\arguments{
\item{X}{Numeric matrix (n_vars x n_time).}

\item{degree}{Maximum total polynomial degree (default 2).}
}
\value{
Lifted matrix including all monomials up to total degree.
For n=2, degree=2: x1, x2, x1^2, x1*x2, x2^2
}
\description{
Lifts state space by adding all polynomial terms including cross-products
up to specified total degree.
}
\examples{
X <- matrix(1:10, nrow = 2)
lifted <- RKoopmanDMD:::lift_poly_cross(X, degree = 2)
# Result: x1, x2, x1^2, x1*x2, x2^2

}
\keyword{internal}
