% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifting.R
\name{lift_polynomial}
\alias{lift_polynomial}
\title{Polynomial Lifting}
\usage{
lift_polynomial(X, degree = 2)
}
\arguments{
\item{X}{Numeric matrix (n_vars x n_time).}

\item{degree}{Maximum polynomial degree (default 2).}
}
\value{
Lifted matrix with dimensions (n_vars * degree) x n_time.
Contains original X and powers X^2, X^3, ..., X^degree.
}
\description{
Lifts state space by adding polynomial terms up to specified degree.
}
\examples{
X <- matrix(1:10, nrow = 2)
lifted <- RKoopmanDMD:::lift_polynomial(X, degree = 3)
# Result has 6 rows: x1, x2, x1^2, x2^2, x1^3, x2^3

}
\keyword{internal}
