% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifting.R
\name{list_lifting_functions}
\alias{list_lifting_functions}
\title{List Available Lifting Functions}
\usage{
list_lifting_functions()
}
\value{
A data frame describing available lifting functions.
}
\description{
Returns information about built-in lifting function options.
}
\examples{
list_lifting_functions()

}
