% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pinv}
\alias{pinv}
\title{Compute pseudo-inverse using SVD}
\usage{
pinv(A, tol = .Machine$double.eps * max(dim(A)) * max(abs(A)))
}
\arguments{
\item{A}{Input matrix}

\item{tol}{Tolerance for singular values}
}
\value{
Pseudo-inverse of A
}
\description{
Computes the Moore-Penrose pseudo-inverse of a matrix.
Used as a fallback when solve() fails.
}
\keyword{internal}
