% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_eigenvalue_magnitudes}
\alias{plot_eigenvalue_magnitudes}
\title{Plot eigenvalue magnitudes (internal)}
\usage{
plot_eigenvalue_magnitudes(magnitudes, ...)
}
\description{
Plot eigenvalue magnitudes (internal)
}
\keyword{internal}
