% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_eigenvalue_spectrum}
\alias{plot_eigenvalue_spectrum}
\title{Plot eigenvalue spectrum (internal)}
\usage{
plot_eigenvalue_spectrum(eigenvalues, ...)
}
\description{
Plot eigenvalue spectrum (internal)
}
\keyword{internal}
