% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_mode_amplitudes}
\alias{plot_mode_amplitudes}
\title{Plot mode amplitudes (internal)}
\usage{
plot_mode_amplitudes(amplitudes, ...)
}
\description{
Plot mode amplitudes (internal)
}
\keyword{internal}
