% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_singular_values}
\alias{plot_singular_values}
\title{Plot singular value decay (internal)}
\usage{
plot_singular_values(singular_values, ...)
}
\description{
Plot singular value decay (internal)
}
\keyword{internal}
