% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gla.R
\name{predict.gla}
\alias{predict.gla}
\title{Predict Using GLA Model}
\usage{
\method{predict}{gla}(object, n_ahead = 10, ...)
}
\arguments{
\item{object}{A \code{"gla"} object.}

\item{n_ahead}{Integer; number of time steps to forecast.}

\item{...}{Additional arguments (unused).}
}
\value{
Matrix of predictions (n_obs x n_ahead).
}
\description{
Forecasts future values using the GLA eigenfunctions and modes.
}
