% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{predict.kdmd}
\alias{predict.kdmd}
\title{Predict from kdmd Object (Deprecated)}
\usage{
\method{predict}{kdmd}(object, data, l = 1, ...)
}
\arguments{
\item{object}{A kdmd object generated with getAMatrix.}

\item{data}{The matrix data for which you wish to predict future columns.}

\item{l}{The length of the predictions (number of columns to predict, default = 1).}

\item{...}{Additional arguments (unused).}
}
\value{
A matrix with l additional columns appended.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This method is deprecated. Use \code{\link[=predict.dmd]{predict.dmd()}} with a dmd object instead.
}
\examples{
\dontrun{
# Old way (deprecated)
m <- matrix(1:30, nrow = 3)
A <- getAMatrix(m)
result <- predict(A, m, l = 5)

# New way
model <- dmd(m)
future <- predict(model, n_ahead = 5)
result <- cbind(m, future)
}

}
\seealso{
\code{\link[=predict.dmd]{predict.dmd()}} for the new API.
}
\keyword{internal}
