% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{predict_dmd}
\alias{predict_dmd}
\title{Legacy predict_dmd Function (Deprecated)}
\usage{
predict_dmd(dmd_result, n_steps, x0 = NULL)
}
\arguments{
\item{dmd_result}{Output from koopman_dmd().}

\item{n_steps}{Number of time steps to predict.}

\item{x0}{Initial condition.}
}
\value{
Matrix of predictions.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Use \code{\link[=predict.dmd]{predict.dmd()}} instead.
}
\keyword{internal}
