% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel.R
\name{predict_matrix_hankel}
\alias{predict_matrix_hankel}
\title{Matrix-based prediction for Hankel-DMD (internal)}
\usage{
predict_matrix_hankel(object, x0, n_ahead)
}
\description{
Matrix-based prediction for Hankel-DMD (internal)
}
\keyword{internal}
