% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel.R
\name{predict_modes_hankel}
\alias{predict_modes_hankel}
\title{Mode-based prediction for Hankel-DMD (internal)}
\usage{
predict_modes_hankel(object, x0, n_ahead)
}
\description{
Mode-based prediction for Hankel-DMD (internal)
}
\keyword{internal}
