% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.dmd}
\alias{print.dmd}
\title{Print Method for DMD Objects}
\usage{
\method{print}{dmd}(x, ...)
}
\arguments{
\item{x}{A \code{"dmd"} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns the input object.
}
\description{
Displays a concise summary of a DMD model.
}
\examples{
t <- seq(0, 10, by = 0.1)
X <- rbind(cos(t), sin(t))
model <- dmd(X)
print(model)

}
