% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gla.R
\name{print.gla}
\alias{print.gla}
\title{Print Method for GLA Objects}
\usage{
\method{print}{gla}(x, ...)
}
\arguments{
\item{x}{A \code{"gla"} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for GLA Objects
}
