% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel.R
\name{print.hankel_dmd}
\alias{print.hankel_dmd}
\title{Print Method for Hankel-DMD Objects}
\usage{
\method{print}{hankel_dmd}(x, ...)
}
\arguments{
\item{x}{A \code{"hankel_dmd"} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for Hankel-DMD Objects
}
