% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.dmd}
\alias{print.summary.dmd}
\title{Print Method for DMD Summary}
\usage{
\method{print}{summary.dmd}(x, ...)
}
\arguments{
\item{x}{A \code{"summary.dmd"} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for DMD Summary
}
