% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_matrix}
\alias{validate_matrix}
\title{Validate input data matrix}
\usage{
validate_matrix(X, min_rows = 2, min_cols = 3)
}
\arguments{
\item{X}{Input data (matrix or data.frame)}

\item{min_rows}{Minimum number of rows required}

\item{min_cols}{Minimum number of columns required}
}
\value{
A numeric matrix
}
\description{
Checks that input is a valid matrix for DMD analysis.
}
\keyword{internal}
